+++ src/index.js
@@ -4,11 +4,15 @@ import { createHashHistory } from "history";
 import { Router, Route, Switch } from "react-router-dom";
 import ReactPixel from "react-facebook-pixel";
 
-import indexRoutes from "routes/index.jsx";
-
 import "assets/scss/material-kit-react.scss?v=1.4.0";
 
 import "react-github-button/assets/style.css";
+
+// pages for this product
+import Components from "views/Components/Components.jsx";
+import LandingPage from "views/LandingPage/LandingPage.jsx";
+import ProfilePage from "views/ProfilePage/ProfilePage.jsx";
+import LoginPage from "views/LoginPage/LoginPage.jsx";
 
 var hist = createHashHistory();
 
@@ -17,7 +21,6 @@ ReactPixel.pageView();
 ReactPixel.fbq("track", "PageView");
 
 hist.listen(location => {
-
   ReactPixel.pageView();
   ReactPixel.fbq("track", "PageView");
 });
@@ -25,9 +28,10 @@ hist.listen(location => {
 ReactDOM.render(
   <Router history={hist}>
     <Switch>
-      {indexRoutes.map((prop, key) => {
-        return <Route path={prop.path} key={key} component={prop.component} />;
-      })}
+      <Route path="/landing-page" component={LandingPage} />
+      <Route path="/profile-page" component={ProfilePage} />
+      <Route path="/login-page" component={LoginPage} />
+      <Route path="/" component={Components} />
     </Switch>
   </Router>,
   document.getElementById("root")